page-optimizer: omit grouped and metrics from data when submittingSettings and testMode UI changes
