+++ includes/admin/page-optimizer/src/app/page-optimizer/components/Header.tsx
@@ -33,9 +33,6 @@ import equal from 'fast-deep-equal/es6/react'
 import UnsavedChanges from "app/page-optimizer/components/footer/unsaved-changes";
 import UrlPreview from "app/page-optimizer/components/footer/url-preview";
 import SaveChanges from "app/page-optimizer/components/footer/save-changes";
-import {Switch} from "components/ui/switch";
-import {getTestModeStatus} from "../../../store/app/appActions";
-import {useToast} from "components/ui/use-toast";
 
 // const Header = ({ url }: { url: string}) => {
 const Header = ({ url }: { url: string}) => {
@@ -68,104 +65,7 @@ const Header = ({ url }: { url: string}) => {
     const {testMode} = useSelector((state: RootState) => state.app);
 
     const dispatch: ThunkDispatch<RootState, unknown, AppAction> = useDispatch();
-    const [timeoutId, setTimeoutId] = useState<NodeJS.Timeout | null>(null);
-    const [localSwitchState, setLocalSwitchState] = useState<boolean>(false);
-    const [previewButton, setPreviewButton]= useState<boolean>(false);
-    const [testModeLoading, setTestModeLoading]= useState<boolean>(false);
 
-    useEffect(() => {
-        dispatch(getTestModeStatus(options, url));
-    }, [dispatch]);
-
-    useEffect(() => {
-        let toastInstance: ReturnType<typeof toast> | undefined;
-        if(settingsMode==='turboMax' && !localSwitchState){
-            toastInstance = toast({
-                description: (
-                    <>
-                        <div className='flex w-full gap-2 text-center items-center'>
-                            <InformationCircleIcon className='w-5 text-orange-600'/>
-                            Do you want to turn on test mode?
-                            <AppButton onClick={e => {
-                                handleSwitchChange(true);
-                            }} variant='outline'>
-                                Yes
-                            </AppButton>
-                            <AppButton onClick={e => {
-                                // Dismiss the toast immediately
-                                if (toastInstance) {
-                                    toastInstance.dismiss();
-                                }
-                            }} variant='outline'>
-                                No
-                            </AppButton>
-                        </div>
-                    </>
-                ),
-            }, 50000);
-
-        }else if(settingsMode!='turboMax'){
-            console.log(settingsMode)
-            if (toastInstance) {
-                console.log("test")
-                toastInstance.dismiss();
-            }
-        } else if(!testMode){
-            setLocalSwitchState(false);
-        }
-    }, [settingsMode]);
-
-    useEffect(() => {
-        if (testMode) {
-            setLocalSwitchState(testMode.status || false);
-            setPreviewButton(true);
-        }
-    }, [testMode]);
-
-    const { toast } = useToast();
-
-    const handleSwitchChange = async (isChecked: boolean) => {
-      //  setLocalSwitchState(isChecked);
-        setTestModeLoading(true);
-
-        if (timeoutId) {
-            clearTimeout(timeoutId);
-        }
-
-        const newTimeoutId = setTimeout(async () => {
-            const result = await dispatch(getTestModeStatus(options, url, String(isChecked)));
-            if (result.success) {
-                setTestModeLoading(false);
-                toast({
-                    description: (
-                        <>
-                            <div className='flex w-full gap-2 text-center'>
-                                Test Mode turned {localSwitchState ? 'off' : 'on'} successfully
-                                <CheckCircleIcon className='w-5 text-green-600'/>
-                            </div>
-                            <div className='flex w-full gap-2 text-center'>
-                                <InformationCircleIcon className='w-5 text-green-600'/>
-                                Test Mode changes are on live
-                            </div>
-                        </>
-                    ),
-                });
-            } else {
-                toast({
-                    description: (
-                        <div className='flex w-full gap-2 text-center'>
-                            Failed to turn on Test mode: {result.error}
-                            <XCircleIcon className='w-5 text-red-600' />
-                        </div>
-                    ),
-                });
-                setLocalSwitchState(false);
-                setTestModeLoading(false);
-            }
-        }, 200);
-        setTimeoutId(newTimeoutId);
-
-    };
 
     return (
         <>
@@ -249,40 +149,6 @@ const Header = ({ url }: { url: string}) => {
                     <>
                     {!error && (
                         <>
-                            <div className="flex overflow-hidden border rounded-2xl shadow">
-
-                                <button
-                                    onClick={() => {
-                                        window.open(options.optimizer_url + '?rapidload_preview_optimization', '_blank');
-                                    }}
-//cursor-not-allowed opacity-50 pointer-events-none
-                                    className={`flex gap-2 items-center text-sm h-10 rounded-[14px] bg-brand-200/80 dark:bg-primary dark:hover:bg-primary/90 px-4 py-2 m-1 pr-5 ${
-                                     revisions.length > 0 
-                                        ? '' : ''}`} data-tour="preview-button">
-
-                                    <ArrowTopRightOnSquareIcon className='w-5 text-gray-500'/>
-                                    Preview
-                                </button>
-
-                                <div
-                                    className="text-sm flex gap-1 justify-end items-center ml-4 mr-4 text-left w-full dark:text-brand-300" data-tour="test-mode">
-                                    <div>{testModeLoading ? ('Test Mode') : ('Test Mode')}</div>
-                                    {testModeLoading ? (
-                                        <Loader className='ml-2 mr-[7px] w-5 animate-spin'/>
-                                    ) : (
-                                        <Switch
-                                            checked={localSwitchState}
-                                            onCheckedChange={(checked) => handleSwitchChange(checked)}
-
-                                        />
-                                    )
-                                    }
-
-
-                                </div>
-                            </div>
-
-
                             <SaveChanges/>
                         </>
                     )}
@@ -306,7 +172,7 @@ const Header = ({ url }: { url: string}) => {
                 </UnsavedChanges>
             </div>
         </header>
-            {localSwitchState && !loading && !showInprogress && !testModeLoading &&
+            {!loading && !showInprogress && testMode &&
                 <motion.div  initial={{ opacity: 0, y: -10 }}
                          animate={{ opacity: 1, y: 0 }}
                          exit={{ opacity: 0, y: -10 }}
