+++ includes/admin/page-optimizer/src/app/page-optimizer/components/audit/Audit.tsx
@@ -228,29 +228,6 @@ const Audit = forwardRef<AuditComponentRef, AuditProps>(({audit, index, actions 
                                     )}
 
 
-                                    {/*{ toggleFiles &&*/}
-                                    {/*    <TooltipText*/}
-                                    {/*        text={filesOrActions ? 'Get step-by-step Help' : 'Learn more about this audit'}*/}
-                                    {/*    >*/}
-                                    {/*        <div onClick={() => setHelpOpen(p => !p)}*/}
-                                    {/*             className={`min-w-[125px] cursor-pointer flex items-center gap-2 pl-4 pr-2 py-1.5 text-sm rounded-2xl dark:hover:bg-brand-800 hover:bg-brand-100 transition-colors ${toggleFiles ? ' dark:bg-brand-900 border ': 'border '}`}>*/}
-
-                                    {/*            {filesOrActions ? (*/}
-                                    {/*                toggleFiles ? 'Titan Help' : 'Show Actions'*/}
-                                    {/*            ) : 'Learn More'}*/}
-
-
-                                    {/*            {*/}
-                                    {/*                (helpOpen) ?*/}
-                                    {/*                    <MinusCircleIcon className='w-6 h-6 dark:text-brand-500 text-brand-900'/>:*/}
-                                    {/*                    <QuestionMarkCircleIcon className='w-6 h-6 dark:text-brand-500 text-brand-900'/>*/}
-                                    {/*            }*/}
-
-                                    {/*        </div>*/}
-                                    {/*    </TooltipText>*/}
-                                    {/*}*/}
-
-
                                     {actions &&
                                         <TooltipText
                                             text={filesOrActions ? 'Show resources and actions' : 'Learn more about this audit'}
