+++ includes/admin/page-optimizer/src/app/page-optimizer/components/performance-widgets/PageSpeedScore.tsx
@@ -12,12 +12,20 @@ import useCommonDispatch from "hooks/useCommonDispatch";
 import {setCommonRootState, setCommonState} from "../../../../store/common/commonActions";
 import {
     Circle, GraduationCapIcon,
-    Hash, History,
+    Hash, History, Monitor,
 } from "lucide-react";
 import SideBarActions from "app/page-optimizer/components/performance-widgets/SideBarActions";
 import xusePerformanceColors from "hooks/usePerformanceColors";
 import AppButton from "components/ui/app-button";
 import Feedback from "app/page-optimizer/components/performance-widgets/Feedback";
+import TooltipText from "components/ui/tooltip-text";
+import {changeReport} from "../../../../store/app/appActions";
+import {ArrowTopRightOnSquareIcon, DevicePhoneMobileIcon, InformationCircleIcon} from "@heroicons/react/24/outline";
+import {getTestModeStatus} from "../../../../store/app/appActions";
+import {useToast} from "components/ui/use-toast";
+import {RootState} from "../../../../store/app/appTypes";
+import {CheckCircleIcon, XCircleIcon} from "@heroicons/react/24/solid";
+import {SettingsStraightLine} from "app/page-optimizer/components/icons/icon-svg";
 
 // const Feedback = React.lazy(() =>
 //     import('app/page-optimizer/components/performance-widgets/Feedback'))
@@ -46,11 +54,121 @@ const PageSpeedScore = ({pagespeed, priority = true }: PageSpeedScoreProps) => {
     const [isCoreWebClicked, setCoreWebIsClicked] = useState(false);
     const [expanded, setExpanded] = useState(false)
 
+
+
     const {data, error, loading, revisions} = useSelector(optimizerData);
     const [performance, setPerformance] = useState<number>(0)
     const [on, setOn] = useState<boolean>(false)
 
     const { dispatch, hoveredMetric, activeMetric} = useCommonDispatch()
+
+    //Test Mode
+    const {options,} = useAppContext();
+    const {settingsMode} = useCommonDispatch();
+    const {testMode} = useSelector((state: RootState) => state.app);
+    const [timeoutId, setTimeoutId] = useState<NodeJS.Timeout | null>(null);
+    const [localSwitchState, setLocalSwitchState] = useState<boolean>(true);
+    const [previewButton, setPreviewButton]= useState<boolean>(false);
+    const [testModeLoading, setTestModeLoading]= useState<boolean>(false);
+    const { toast } = useToast();
+
+    let url = options?.optimizer_url;
+
+    useEffect(() => {
+        dispatch(getTestModeStatus(options, url));
+    }, [dispatch]);
+
+    useEffect(() => {
+        let toastInstance: ReturnType<typeof toast> | undefined;
+        if(settingsMode==='turboMax' && !localSwitchState){
+            toastInstance = toast({
+                description: (
+                    <>
+                        <div className='flex w-full gap-2 text-center items-center'>
+                            <InformationCircleIcon className='w-5 text-orange-600'/>
+                            Do you want to turn on test mode?
+                            <AppButton onClick={e => {
+                                handleSwitchChange(true);
+                            }} variant='outline'>
+                                Yes
+                            </AppButton>
+                            <AppButton onClick={e => {
+                                // Dismiss the toast immediately
+                                if (toastInstance) {
+                                    toastInstance.dismiss();
+                                }
+                            }} variant='outline'>
+                                No
+                            </AppButton>
+                        </div>
+                    </>
+                ),
+            }, 50000);
+
+        }else if(settingsMode!='turboMax'){
+            console.log(settingsMode)
+            if (toastInstance) {
+                console.log("test")
+                toastInstance.dismiss();
+            }
+        } else if(!testMode){
+           // setLocalSwitchState(false);
+        }
+    }, [settingsMode]);
+
+    useEffect(() => {
+        if (testMode) {
+            setLocalSwitchState(testMode.status || false);
+            setPreviewButton(true);
+        }
+    }, [testMode]);
+
+
+
+    const handleSwitchChange = async (isChecked: boolean) => {
+
+        setLocalSwitchState(isChecked);
+        setTestModeLoading(true);
+
+        if (timeoutId) {
+            clearTimeout(timeoutId);
+        }
+
+        const newTimeoutId = setTimeout(async () => {
+            const result = await dispatch(getTestModeStatus(options, url, String(isChecked)));
+            if (result.success) {
+                setTestModeLoading(false);
+                toast({
+                    description: (
+                        <>
+                            <div className='flex w-full gap-2 text-center'>
+                                Test Mode turned {localSwitchState ? 'off' : 'on'} successfully
+                                <CheckCircleIcon className='w-5 text-green-600'/>
+                            </div>
+                            <div className='flex w-full gap-2 text-center'>
+                                <InformationCircleIcon className='w-5 text-green-600'/>
+                                Test Mode changes are on live
+                            </div>
+                        </>
+                    ),
+                });
+            } else {
+                toast({
+                    description: (
+                        <div className='flex w-full gap-2 text-center'>
+                            Failed to turn on Test mode: {result.error}
+                            <XCircleIcon className='w-5 text-red-600' />
+                        </div>
+                    ),
+                });
+                setLocalSwitchState(false);
+                setTestModeLoading(false);
+            }
+        }, 200);
+        setTimeoutId(newTimeoutId);
+
+    };
+
 
     const handleCoreWebClick = useCallback(() => {
         setCoreWebIsClicked(!isCoreWebClicked);
@@ -103,12 +221,57 @@ const PageSpeedScore = ({pagespeed, priority = true }: PageSpeedScoreProps) => {
 
     return <>
         {/*min-w-[310px]*/}
+
         <div className='w-full flex flex-col gap-4'>
+            <div className='flex gap-2 justify-center'>
+                <div className='w-fit'>
+                    <div data-tour='test-mode'
+                         className='select-none relative flex dark:bg-brand-800 py-0.5 px-0.5 rounded-2xl cursor-pointer bg-brand-0'>
+                        <div className={cn(
+                            'absolute translate-x-0 left-0.5 w-[70px] rounded-[14px] -z-1 duration-300 h-[44px] text-sm flex flex-col gap-2 px-3 py-2.5 font-medium dark:bg-brand-950 bg-brand-200/80',
+                            localSwitchState && 'w-[118px] -translate-x-1 left-[40%] bg-amber-500/80'
+                        )}>
+                        </div>
+
+                        <div
+                            onClick={() => setLocalSwitchState(false)}
+                            className={`relative z-1 items-center text-sm flex gap-2 px-3 py-2.5 font-medium rounded-2xl ${localSwitchState ? 'text-brand-500' : ''}`}
+                        >
+                            <Circle
+                                className={cn(`w-1.5 stroke-0 ${localSwitchState ? '' : 'fill-green-600'} animate-ping absolute inline-flex opacity-75`)}/>
+                            <Circle
+                                className={cn(`w-1.5 stroke-0 ${localSwitchState ? '' : 'fill-green-600'} relative inline-flex`)}/>
+                            Live
+                        </div>
+
+                        <div
+                            onClick={() => setLocalSwitchState(true)}
+                            className={`relative justify-center items-center z-1 text-sm flex pl-8 pr-6 py-2.5 whitespace-nowrap font-medium rounded-2xl ${localSwitchState ? 'text-brand-0' : ''}`}
+                        >
+                            Test Mode
+                        </div>
+                    </div>
+                </div>
+                <TooltipText text="Preview">
+                <button
+                    onClick={() => {
+                        window.open(options.optimizer_url + '?rapidload_preview_optimization', '_blank');
+                    }}
+                    className={`flex gap-2 items-center text-sm h-12 rounded-[14px] bg-brand-0 dark:bg-primary dark:hover:bg-primary/90 px-4 py-2 pr-3.5 ${
+                        revisions.length > 0
+                            ? '' : ''}`} data-tour="preview-button">
+
+                    <ArrowTopRightOnSquareIcon className='w-5 text-gray-500'/>
+                </button>
+                </TooltipText>
+            </div>
+            <SettingsStraightLine/>
             <Card data-tour='speed-insights'
-                className={cn(
-                    'overflow-hidden border border-transparent flex flex-col sm:flex-row lg:flex-col justify-around',
-                    expanded && 'border-brand-200 dark:border-brand-800'
-                )}>
+                  className={cn(
+                      'overflow-hidden border border-transparent flex flex-col sm:flex-row lg:flex-col justify-around',
+                      expanded && 'border-brand-200 dark:border-brand-800'
+                  )}>
+
                 <div
                     className="content flex w-full sm:w-1/2 lg:w-full flex-col justify-center items-center gap-3 px-4 lg:px-4 lg:pb-0 xl:px-8 py-2.5">
 
@@ -143,7 +306,8 @@ const PageSpeedScore = ({pagespeed, priority = true }: PageSpeedScoreProps) => {
                             Values are estimated and may vary with Google Page Speed Insights.
                         </div>
                     </div>
-                    <div className="flex justify-around text-sm gap-4 font-normal w-full mb-5 text-brand-700 dark:text-brand-300">
+                    <div
+                        className="flex justify-around text-sm gap-4 font-normal w-full mb-5 text-brand-700 dark:text-brand-300">
                         <div className="flex lg:flex-col xl:flex-row items-center gap-1">
                             <PerformanceIcons icon={'fail'}/>
                             0-49
@@ -161,18 +325,20 @@ const PageSpeedScore = ({pagespeed, priority = true }: PageSpeedScoreProps) => {
 
                 <AppButton
                     onClick={e => setExpanded(p => !p)}
-                    variant='outline' className='select-none border-none bg-transparent hover:bg-transparent text-center text-xs text-brand-600 py-2' data-tour="expand-metrics">
-                    {expanded ? 'Collapse' : 'Expand' } Metrics
+                    variant='outline'
+                    className='select-none border-none bg-transparent hover:bg-transparent text-center text-xs text-brand-600 py-2'
+                    data-tour="expand-metrics">
+                    {expanded ? 'Collapse' : 'Expand'} Metrics
                 </AppButton>
 
                 {(data?.metrics && !expanded) && (
                     <>
                         <div className='flex justify-around mb-3 px-2'
-                             onMouseLeave={() => dispatch(setCommonState('hoveredMetric',null))}
+                             onMouseLeave={() => dispatch(setCommonState('hoveredMetric', null))}
                         >
                             {data.metrics.map(metric => (
                                 <div key={metric.id}
-                                     onMouseEnter={() => dispatch(setCommonState('hoveredMetric',metric))}
+                                     onMouseEnter={() => dispatch(setCommonState('hoveredMetric', metric))}
 
                                      className='text-xs border text-center flex flex-col
                              gap-0.5 px-3 py-2 bg-brand-100/20 hover:bg-brand-100 cursor-default rounded-[14px]'>
@@ -186,14 +352,13 @@ const PageSpeedScore = ({pagespeed, priority = true }: PageSpeedScoreProps) => {
                 )}
 
 
-
-                {(data?.metrics && expanded ) && (
+                {(data?.metrics && expanded) && (
                     <div className={cn(
                         'sticky top-0 w-full sm:w-1/2 lg:w-full border-l lg:border-l-0'
                     )
                     } data-tour='metrics'>
                         <div
-                            onClick={e => dispatch(setCommonState('activeMetric', null)) }
+                            onClick={e => dispatch(setCommonState('activeMetric', null))}
                             className={cn(
                                 'flex gap-3 items-center font-medium dark:hover:bg-brand-900/70  px-6 py-3 border-b lg:border-b-0 lg:border-t cursor-pointer text-sm',
                                 !activeMetric && 'bg-brand-100/80 dark:bg-brand-900/80 '
